fix issue #77

+++ stdplugins/google.py
@@ -5,16 +5,14 @@ Available Commands:
 .google reverse search"""
 
 import asyncio
+import aiohttp
 import os
-import requests
+import shutil
+import time
 from bs4 import BeautifulSoup
 from datetime import datetime
-from google_images_download import google_images_download
+from telethon.utils import guess_extension
 from uniborg.util import admin_cmd
-
-
-def progress(current, total):
-    logger.info("Downloaded {} of {}\nCompleted {}".format(current, total, (current / total) * 100))
 
 
 @borg.on(admin_cmd(pattern="google search (.*)"))
@@ -27,7 +25,9 @@ async def _(event):
     input_str = event.pattern_match.group(1) # + " -inurl:(htm|html|php|pls|txt) intitle:index.of \"last modified\" (mkv|mp4|avi|epub|pdf|mp3)"
     input_url = "https://bots.shrimadhavuk.me/search/?q={}".format(input_str)
     headers = {"USER-AGENT": "UniBorg"}
-    response = requests.get(input_url, headers=headers).json()
+    async with aiohttp.ClientSession() as requests:
+        reponse = await requests.get(input_url, headers=headers)
+        response = await reponse.json()
     output_str = " "
     for result in response["results"]:
         text = result.get("title")
@@ -49,36 +49,55 @@ async def _(event):
     start = datetime.now()
     await event.edit("Processing ...")
     input_str = event.pattern_match.group(1)
-    response = google_images_download.googleimagesdownload()
-    if not os.path.isdir(Config.TMP_DOWNLOAD_DIRECTORY):
-        os.makedirs(Config.TMP_DOWNLOAD_DIRECTORY)
-    arguments = {
-        "keywords": input_str,
-        "limit": Config.TG_GLOBAL_ALBUM_LIMIT,
-        "format": "jpg",
-        "delay": 1,
-        "safe_search": True,
-        "output_directory": Config.TMP_DOWNLOAD_DIRECTORY
-    }
-    paths = response.download(arguments)
-    logger.info(paths)
-    lst = paths[0].get(input_str)
-    if len(lst) == 0:
-        await event.delete()
-        return
-    await borg.send_file(
-        event.chat_id,
-        lst,
-        caption=input_str,
-        reply_to=event.message.id,
-        progress_callback=progress
+    work_dir = os.path.join(
+        Config.TMP_DOWNLOAD_DIRECTORY,
+        input_str
     )
-    logger.info(lst)
-    for each_file in lst:
+    if not os.path.isdir(work_dir):
+        os.makedirs(work_dir)
+    input_url = "https://bots.shrimadhavuk.me/search/?u={}".format(input_str)
+    headers = {"USER-AGENT": "UniBorg"}
+    async with aiohttp.ClientSession() as requests:
+        reponse = await requests.get(input_url, headers=headers)
+        response = await reponse.json()
+    url_lst = []
+    cap_lst = []
+    async with aiohttp.ClientSession() as requests:
+        for result in response["results"]:
+            caption = result.get("description")
+            image_url = result.get("url")
+            image_req_set = await requests.get(image_url)
+            image_file_name = str(time.time()) + "" + guess_extension(
+                image_req_set.headers.get("Content-Type")
+            )
+            image_save_path = os.path.join(
+                work_dir,
+                image_file_name
+            )
+            with open(image_save_path, "wb") as f_d:
+                f_d.write(await image_req_set.read())
+            url_lst.append(image_save_path)
+            cap_lst.append(caption)
+    if len(url_lst) == 0:
+        await event.edit(f"no results found for {input_str}")
+        return
+    if len(url_lst) != len(cap_lst):
+        await event.edit("search api broken :(")
+        return
+    await event.reply(
+        cap_lst,
+        file=url_lst,
+        parse_mode="html"
+    )
+    for each_file in url_lst:
         os.remove(each_file)
+    shutil.rmtree(work_dir, ignore_errors=True)
     end = datetime.now()
     ms = (end - start).seconds
-    await event.edit("searched Google for {} in {} seconds.".format(input_str, ms), link_preview=False)
+    await event.edit(
+        f"searched Google for {input_str} in {ms} seconds.",
+        link_preview=False
+    )
     await asyncio.sleep(5)
     await event.delete()
 
@@ -90,13 +109,15 @@ async def _(event):
     start = datetime.now()
     BASE_URL = "http://www.google.com"
     OUTPUT_STR = "Reply to an image to do Google Reverse Search"
-    if event.reply_to_msg_id:
-        await event.edit("Pre Processing Media")
-        previous_message = await event.get_reply_message()
-        previous_message_text = previous_message.message
+    if not event.reply_to_msg_id:
+        await event.edit("reply to any media, please -_-")
+        return
+    await event.edit("Pre Processing Media")
+    previous_message = await event.get_reply_message()
+    previous_message_text = previous_message.message
+    async with aiohttp.ClientSession() as requests:
         if previous_message.media:
-            downloaded_file_name = await borg.download_media(
-                previous_message,
+            downloaded_file_name = await previous_message.download_media(
                 Config.TMP_DOWNLOAD_DIRECTORY
             )
             SEARCH_URL = "{}/searchbyimage/upload".format(BASE_URL)
@@ -105,21 +126,28 @@ async def _(event):
                 "image_content": ""
             }
             # https://stackoverflow.com/a/28792943/4723940
-            google_rs_response = requests.post(SEARCH_URL, files=multipart, allow_redirects=False)
+            google_rs_response = await requests.post(
+                SEARCH_URL,
+                data=multipart,
+                allow_redirects=False
+            )
             the_location = google_rs_response.headers.get("Location")
             os.remove(downloaded_file_name)
         else:
             previous_message_text = previous_message.message
             SEARCH_URL = "{}/searchbyimage?image_url={}"
             request_url = SEARCH_URL.format(BASE_URL, previous_message_text)
-            google_rs_response = requests.get(request_url, allow_redirects=False)
+            google_rs_response = await requests.get(request_url, allow_redirects=False)
             the_location = google_rs_response.headers.get("Location")
         await event.edit("Found Google Result. Pouring some soup on it!")
         headers = {
             "User-Agent": "Mozilla/5.0 (X11; Linux x86_64; rv:58.0) Gecko/20100101 Firefox/58.0"
         }
-        response = requests.get(the_location, headers=headers)
-        soup = BeautifulSoup(response.text, "html.parser")
+        response = await requests.get(the_location, headers=headers)
+        soup = BeautifulSoup(
+            await response.text(),
+            "html.parser"
+        )
         # document.getElementsByClassName("r5a77d"): PRS
         prs_div = soup.find_all("div", {"class": "r5a77d"})[0]
         prs_anchor_element = prs_div.find("a")
@@ -130,8 +158,7 @@ async def _(event):
         img_size = img_size_div.find_all("div")
         end = datetime.now()
         ms = (end - start).seconds
-        OUTPUT_STR = """{img_size}
-**Possible Related Search**: <a href="{prs_url}">{prs_text}</a>
-
-More Info: Open this <a href="{the_location}">Link</a> in {ms} seconds""".format(**locals())
-    await event.edit(OUTPUT_STR, parse_mode="HTML", link_preview=False)
+        out_rts = f"{img_size}\n"
+        out_rts += f"**Possible Related Search**: <a href='{prs_url}'>{prs_text}</a>\n\n"
+        out_rts += f"More Info: Open this <a href='{the_location}''>Link</a> in {ms} seconds"
+        await event.edit(out_rts, parse_mode="HTML", link_preview=False)
